Added obstacle test cases
Tested Flat image loading and Unlimited distance to Obstacle
From the testing, we learned that there is a strange y-offset by 1,
and the detection algorithm is quite sensitive.

Following this, we fixed the y-offset by referencing the top left
of the image as the origin, and making y pointing downwards be
considered positive. Test cases were then added for distance-
limited, and all test cases now pass.