Allow variable sample points per lidar scan
Previously each lidar scan consisted of a fixed number of sample points,
distributed evenly across 2pi. The data was simple an array of
distances, since the angle could be calculated from the index. Now, data
is stored as a (pointer to a) vector of SamplePoint, each of which
stores its distance and angle. So the number of sample points can vary
between scans (due to minor changes in lidar scanner motor speed for
example), and are not assumed to be perfectly evenly distributed.