Use piecewise potential function
Previously used only a quadratic potential, now the potential function
switches to a linear/conical potential when beyond some distance from
the goal. This prevents the robot from speeding up and slowing down
repeatedly as it switches between waypoints, while still allowing the
robot to slow down when it gets near to the final waypoint.